+++ app/src/main/java/coljamkop/tabtest/Pickers/TimePickerFragment.java
@@ -1,4 +1,4 @@
-package coljamkop.tabtest;
+package coljamkop.tabtest.Pickers;
 
 import android.app.Dialog;
 import android.app.TimePickerDialog;
