+++ app/src/main/java/coljamkop/tabtest/RecyclerViewAdapters/MyAppointmentRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package coljamkop.tabtest;
+package coljamkop.tabtest.RecyclerViewAdapters;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -8,14 +8,15 @@ import android.widget.TextView;
 
 import java.util.List;
 
-import coljamkop.tabtest.AppointmentViewFragment.OnAppointmentListFragmentInteractionListener;
+import coljamkop.tabtest.ViewFragments.AppointmentViewFragment.OnAppointmentListFragmentInteractionListener;
 import coljamkop.tabtest.Content.FamilyContent.Appointment;
+import coljamkop.tabtest.R;
 
 import static coljamkop.tabtest.Content.FamilyContent.Family;
 
 /**
  * {@link RecyclerView.Adapter} that can display a {@link Appointment} and makes a call to the
- * specified {@link AppointmentViewFragment.OnAppointmentListFragmentInteractionListener}.
+ * specified {@link OnAppointmentListFragmentInteractionListener}.
  * TODO: Replace the implementation with code for your data type.
  */
 public class MyAppointmentRecyclerViewAdapter extends RecyclerView.Adapter<MyAppointmentRecyclerViewAdapter.ViewHolder> {
