+++ app/src/main/res/layout/fragment_appointment_list.xml
@@ -15,7 +15,7 @@ android:layout_height="match_parent">
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".AppointmentViewFragment"
+        tools:context=".ViewFragments.AppointmentViewFragment"
         tools:listitem="@layout/fragment_appointment" />
 
     <android.support.design.widget.FloatingActionButton
